calc/ui: slight refactor to allow for monster vs player calculations
