+++ src/app/components/results/NPCVersusPlayerResults.tsx
@@ -1,14 +1,33 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import { observer } from 'mobx-react-lite';
 import { useStore } from '@/state';
 import SectionAccordion from '@/app/components/generic/SectionAccordion';
 import defence from '@/public/img/bonuses/defence.png';
+import attack from '@/public/img/bonuses/attack.png';
+import ranged from '@/public/img/bonuses/ranged.png';
+import magic from '@/public/img/bonuses/magic.png';
 import LazyImage from '@/app/components/generic/LazyImage';
 import { IconAlertTriangle } from '@tabler/icons-react';
 
 const NPCVersusPlayerResults: React.FC = observer(() => {
   const store = useStore();
+  const { monster } = store;
   const { showNPCVersusPlayerResults } = store.prefs;
+
+  const renderMonsterStyleImage = useMemo(() => {
+    if (['slash', 'stab', 'crush'].includes(monster.style || '')) {
+      return <img src={attack.src} className="w-5" alt="Melee" />;
+    }
+    if (monster.style === 'ranged') {
+      return <img src={ranged.src} className="w-5" alt="Ranged" />;
+    }
+    if (monster.style === 'magic') {
+      return <img src={magic.src} className="w-5" alt="Magic" />;
+    }
+    return null;
+  }, [monster.style]);
+
+  const isNonStandard = useMemo(() => !['slash', 'crush', 'stab', 'magic', 'ranged'].includes(monster.style || ''), [monster.style]);
 
   return (
     <SectionAccordion
@@ -23,14 +42,28 @@ const NPCVersusPlayerResults: React.FC = observer(() => {
         </div>
       )}
     >
-      <div
-        className="w-full bg-orange-500 text-white px-4 py-1 text-sm border-b border-orange-400 flex items-center gap-2"
-      >
-        <IconAlertTriangle className="text-orange-200" />
-        This monster has non-standard behaviour. This section could be inaccurate.
-      </div>
-      <div className="px-6 py-4 text-white">
-        Lorem ipsum
+      {isNonStandard && (
+        <div
+          className="w-full bg-orange-500 text-white px-4 py-1 text-sm border-b border-orange-400 flex items-center gap-2"
+        >
+          <IconAlertTriangle className="text-orange-200" />
+          This monster has non-standard behaviour. This section could be inaccurate.
+        </div>
+      )}
+      <div className="px-6 py-4 text-white flex gap-4">
+        <div className="border-l-2 bg-dark-400 px-3 py-1 rounded-r border-blue-300 flex flex-col text-sm">
+          <div className="flex gap-1">
+            <span className="text-gray-300">Monster style:</span>
+            <div className="flex gap-1">
+              {renderMonsterStyleImage}
+              <span className="capitalize">{monster.style}</span>
+            </div>
+          </div>
+          <div className="flex gap-2 items-center">
+            <span className="text-gray-300">Monster speed:</span>
+            {monster.speed}
+          </div>
+        </div>
       </div>
     </SectionAccordion>
   );
